+++ lua/obsidian/init.lua
@@ -9,6 +9,7 @@ obsidian.VERSION = "1.12.0"
 obsidian.completion = require "obsidian.completion"
 obsidian.note = require "obsidian.note"
 obsidian.util = require "obsidian.util"
+obsidian.mapping = require "obsidian.mapping"
 
 ---@class obsidian.Client
 ---@field dir Path
@@ -92,21 +93,21 @@ obsidian.setup = function(opts)
       cmp.setup.buffer { sources = sources }
     end
 
-    -- Keybindings...
-    -- 'gf' passthrough
-    if opts.mappings.gf_passthrough then
-      local gf_set_by = vim.fn.mapcheck("gf", "n")
-      if gf_set_by == "" then
-        -- 'gf' has not been overridden by user or other plugin so we can safely set it.
-        vim.keymap.set("n", "gf", function()
-          if require("obsidian").util.cursor_on_markdown_link() then
-            return "<cmd>ObsidianFollowLink<CR>"
-          else
-            return "gf"
-          end
-        end, { noremap = false, expr = true, buffer = true })
-      elseif not string.find(gf_set_by, "obsidian") then
-        echo.warn("Obsidian cannot override 'gf' keybinding since 'gf' is already set by " .. gf_set_by, opts.log_level)
+    -- Mappings...
+    for mapping_keys, mapping_config in pairs(opts.mappings) do
+      local mapping_set_by = vim.fn.mapcheck(mapping_keys, "n")
+      if mapping_set_by == "" then
+        vim.keymap.set("n", mapping_keys, mapping_config.action, mapping_config.opts)
+      elseif not string.find(mapping_set_by, "obsidian") then
+        echo.warn(
+          "Obsidian cannot override '"
+            .. mapping_keys
+            .. "' keybinding since '"
+            .. mapping_keys
+            .. "' has been set by "
+            .. mapping_set_by,
+          opts.log_level
+        )
       end
     end
   end