+++ CHANGELOG.md
@@ -9,12 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added option `prepend_note_id` to allow disabling id generation for new notes
+- Added option `prepend_note_id` to allow disabling id generation for new notes.
+- Added `mappings` configuration field.
 
 ### Changed
 
 - (internal) Refactored daily note creation.
-- obsidian.nvim will now automatically enable the 'gf' passthrough keybinding within your vault unless the 'gf' keybinding has already been overridden you or another plugin or you set 'mappings.gf_passthrough = false' in your obsidian.nvim config.
+- obsidian.nvim will now automatically enable the 'gf' passthrough keybinding within your vault unless the 'gf' keybinding has already been overridden by you or another plugin or you override the 'mappings' configuration field.
 
 ## [v1.12.0](https://github.com/epwalsh/obsidian.nvim/releases/tag/v1.12.0) - 2023-07-15